requirements.txt
sklearn
pandas
numpy
unidecode
tqdm
matplotlib
seaborn
pyspellchecker
nltk